NaN problem fix
The problem was the sharpening function in memory.py which had a division with no guarantee that the denominator is not zero.

I have tested it by starting to learn from a checkpoint which was very close to outputting NaNs (it always happened in a couple of hundred iterations) and with this fix it worked fine for 300k more iterations.

Of course the result of the loss function changed a little at first, so this is not a 100% guarantee, but I'm confident this patch removes one possibility for the NaN problem.